restore: reorganize progress bar code
The structure is now much more similar to that of the backup command.
