+++ internal/ui/restore/progress.go
@@ -1,11 +1,9 @@
 package restore
 
 import (
-	"fmt"
 	"sync"
 	"time"
 
-	"github.com/restic/restic/internal/ui"
 	"github.com/restic/restic/internal/ui/progress"
 )
 
@@ -26,6 +24,11 @@ type Progress struct {
 type progressInfoEntry struct {
 	bytesWritten uint64
 	bytesTotal   uint64
+}
+
+type term interface {
+	Print(line string)
+	SetStatus(lines []string)
 }
 
 type ProgressPrinter interface {
@@ -84,48 +87,4 @@ func (p *Progress) AddProgress(name string, bytesWrittenPortion uint64, bytesTot
 
 func (p *Progress) Finish() {
 	p.updater.Done()
-}
-
-type term interface {
-	Print(line string)
-	SetStatus(lines []string)
-}
-
-type textPrinter struct {
-	terminal term
-}
-
-func NewProgressPrinter(terminal term) ProgressPrinter {
-	return &textPrinter{
-		terminal: terminal,
-	}
-}
-
-func (t *textPrinter) Update(filesFinished, filesTotal, allBytesWritten, allBytesTotal uint64, duration time.Duration) {
-	timeLeft := ui.FormatDuration(duration)
-	formattedAllBytesWritten := ui.FormatBytes(allBytesWritten)
-	formattedAllBytesTotal := ui.FormatBytes(allBytesTotal)
-	allPercent := ui.FormatPercent(allBytesWritten, allBytesTotal)
-	progress := fmt.Sprintf("[%s] %s  %v files %s, total %v files %v",
-		timeLeft, allPercent, filesFinished, formattedAllBytesWritten, filesTotal, formattedAllBytesTotal)
-
-	t.terminal.SetStatus([]string{progress})
-}
-
-func (t *textPrinter) Finish(filesFinished, filesTotal, allBytesWritten, allBytesTotal uint64, duration time.Duration) {
-	t.terminal.SetStatus([]string{})
-
-	timeLeft := ui.FormatDuration(duration)
-	formattedAllBytesTotal := ui.FormatBytes(allBytesTotal)
-
-	var summary string
-	if filesFinished == filesTotal && allBytesWritten == allBytesTotal {
-		summary = fmt.Sprintf("Summary: Restored %d Files (%s) in %s", filesTotal, formattedAllBytesTotal, timeLeft)
-	} else {
-		formattedAllBytesWritten := ui.FormatBytes(allBytesWritten)
-		summary = fmt.Sprintf("Summary: Restored %d / %d Files (%s / %s) in %s",
-			filesFinished, filesTotal, formattedAllBytesWritten, formattedAllBytesTotal, timeLeft)
-	}
-
-	t.terminal.Print(summary)
 }
